Move: move Sibala, Player, Sibala to production
